Moved test to a separate folder
